Move some classes to storage package
