+++ src/main/java/goober/Goober.java
@@ -1,5 +1,7 @@
 package goober;
 
+import goober.storage.SaveData;
+import goober.storage.Storage;
 import goober.task.Deadline;
 import goober.task.Event;
 import goober.task.Task;
