+++ src/test/java/goober/ParserTest.java
@@ -1,0 +1,19 @@
+package goober;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class ParserTest {
+    @Test
+    public void dummyTest(){
+        assertEquals(2, 2);
+    }
+
+    @Test
+    public void anotherDummyTest(){
+        assertEquals(4, 4);
+    }
+}
+
+
